add middlware for pormpt and user as context

+++ models/prompt.go
@@ -4,7 +4,7 @@ import "time"
 
 type Prompt struct {
 	ID        int       `json:"id"`
-	UserID    float64   `json:"user_id"`
+	UserID    int64     `json:"user_id"`
 	Title     string    `json:"title"`
 	Body      string    `json:"body"`
 	CreatedAt time.Time `json:"created_at"`
